requirements.txt
openai>=1.0.0
anthropic>=0.18.0
chromadb>=0.4.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyGithub>=2.1.1
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
aiohttp>=3.9.0
